Create Practical 12
Create a Scala Application, which would include different user-defined functions for given tasks:
Task-01. WAP in Scala to implement var args in a trait class, so the sub class can implement trait show method which would accept variable length parameters. You must define show method in such a way so it can accept any number of parameters.

Task-02. WAP in Scala to accept a number and if the number is positive then print the square of that number but if the number is negative then it should return "Negative Number”.