Create Practical 13
Write Scala programs for given tasks in the Description.


Task-01. WAP in Scala, in which you read a month from console, and if the entered month is October, November, December, January, February then it would print it is winter now else it should print "It is $Month, and Winter will come again!”

Task-02. WAP in Scala in which you read a string from a console, and from this string find how many times "can" or "Can" word is used and also at the end your program should be able to replace all "can" with "could" even if the string would not make any sense anymore.